requirements.txt
pandas
openpyxl
boto3
apache-airflow-providers-amazon